Simply assignment titles and home display
All the labs and problem set titles had
"CSci 1301: " in front of them, which was
unnecessary clutter, so I removed all that.
They also had things like "CSci 1301: Lab 1"
at the top of each page, which was 
duplicated by the title, so I deleted that,
along with the due dates, which will live
on Canvas.

Lastly, I updated `index.html` so that the
display of the problem sets and labs are
in an unordered list instead of a sequence
of big `<h2>` elements.