Create separate lab & problem set pages
This extracts the loop that lists all the labs & the loop that lists the 
problem sets out into separate includes, and then uses that to create 
pages that list all the labs and all the problem sets, respectively. 

I also changed `categories` to `tags` for a couple of labs and problem 
sets. They don't get included by Jekyll, which apparently only does tags 
and categories for posts. I'm not going to homebrew a replacement right 
now, but it can be done if I want to do it later.

I removed the one "demo" post so the posts section no long shows up on 
the front page. I doubt I'll uses posts, as I'll just do that kind of 
thing in Canvas instead.